Support for Typescript typings.
