+++ .babelrc
@@ -1,7 +1,8 @@
 {
   "presets": [
-    ["@babel/preset-env", { "modules": "umd" }],
-    "@babel/preset-react"
+    "@babel/preset-env",
+    "@babel/preset-react",
+    "@babel/preset-typescript"
   ],
   "plugins": [
     "@babel/plugin-proposal-class-properties",
