+++ .browserslistrc
@@ -1,0 +1,9 @@
+[production]
+>0.2%
+not dead
+not op_mini all
+
+[development]
+last 1 chrome version
+last 1 firefox version
+last 1 safari version
