+++ src/components/SideNavMenuItem.jsx
@@ -1,0 +1,17 @@
+import React from "react";
+import PropTypes from "prop-types";
+
+export function SideNavMenuItem(props) {
+    return (
+        <SideNavItem
+            {...props}
+            className={`second-level-nav-item ${props.className || ""}`}
+        ></SideNavItem>
+    );
+}
+
+SideNavMenuItem.propTypes = {
+    className: PropTypes.string
+};
+
+export default SideNavMenuItem;
