+++ src/components/index.js
@@ -1,25 +1,24 @@
-export { default as Accordion } from "./Accordion";
-export { default as AccordionSet } from "./AccordionSet";
-export { default as Brand } from "./Brand";
-export { default as Breadcrumb, BreadcrumbItem } from "./Breadcrumb";
-export { default as Button } from "./Button";
-export { default as Callout } from "./Callout";
-export { default as Card } from "./Card";
-export { default as Dropdown } from "./Dropdown";
-export { default as DropdownItem } from "./DropdownItem";
-export { default as Hero } from "./Hero";
-export { default as Masthead } from "./Masthead";
-export { default as Footer } from "./Footer";
-export { default as MainNav } from "./MainNav";
-export { default as Tab } from "./Tab";
-export {
-  default as SideNav,
-  SideNavItem,
-  SideNavMenu,
-  SideNavMenuItem,
-} from "./SideNav";
-export { default as Container } from "./Container";
-export { default as Section } from "./Section";
-export { default as Row } from "./Row";
-export { default as Col } from "./Col";
-export { default as Notification } from "./Notification";
+export * from "./Accordion";
+export * from "./AccordionSet";
+export * from "./Brand";
+export * from "./Breadcrumb";
+export * from "./BreadcrumbItem";
+export * from "./Button";
+export * from "./Callout";
+export * from "./Card";
+export * from "./Dropdown";
+export * from "./DropdownItem";
+export * from "./Hero";
+export * from "./Masthead";
+export * from "./Footer";
+export * from "./MainNav";
+export * from "./Tab";
+export * from "./SideNav";
+export * from "./SideNavItem";
+export * from "./SideNavMenu";
+export * from "./SideNavMenuItem";
+export * from "./Container";
+export * from "./Section";
+export * from "./Row";
+export * from "./Col";
+export * from "./Notification";
